LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/scroll-timeline-responsiveness-from-endpoint-expected.txt


FAIL Test that the scroll animation is still responsive after moving from 100% assert_equals: 'actual' unit type must be 'percent' for "Timeline time when animation is ready" expected (string) "percent" but got (undefined) undefined
